requirements.txt
django
psycopg[binary]
django-ninja

